refactor(plugins): Breakup multi-plugin files
