+++ lua/plugins/autopair.lua
@@ -2,29 +2,49 @@ return {
   "windwp/nvim-autopairs",
   event = { "InsertEnter" },
   dependencies = {
-    "hrsh7th/nvim-cmp",
+    "saghen/blink.cmp",
   },
   config = function()
-    -- import nvim-autopairs
     local autopairs = require("nvim-autopairs")
 
-    -- configure autopairs
     autopairs.setup({
-      check_ts = true, -- enable treesitter
+      check_ts = true,
       ts_config = {
-        lua = { "string" }, -- don't add pairs in lua string treesitter nodes
-        javascript = { "template_string" }, -- don't add pairs in javscript template_string treesitter nodes
-        java = false, -- don't check treesitter on java
+        lua = { "string" },
+        javascript = { "template_string" },
+        java = false,
       },
     })
 
-    -- import nvim-autopairs completion functionality
-    local cmp_autopairs = require("nvim-autopairs.completion.cmp")
-
-    -- import nvim-cmp plugin (completions plugin)
-    local cmp = require("cmp")
-
-    -- make autopairs and completion work together
-    cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done())
+    -- blink.cmp integration
+    local blink = require("blink.cmp")
+    if blink and blink.setup then
+      autopairs.setup({
+        map_char = {
+          all = "(",
+          tex = "{",
+        },
+        enable_check_bracket_line = false,
+        check_ts = true,
+        ts_config = {
+          lua = { "string", "source" },
+          javascript = { "string", "template_string" },
+          java = false,
+        },
+        disable_filetype = { "TelescopePrompt", "spectre_panel" },
+        disable_in_macro = true,
+        disable_in_visualblock = false,
+        disable_in_replace_mode = true,
+        ignored_next_char = string.gsub([[ [%w%%%'%[%"%.] ]], "%s+", ""),
+        enable_moveright = true,
+        enable_afterquote = true,
+        enable_bracket_in_quote = true,
+        enable_abbr = false,
+        break_undo = true,
+        map_bs = true,
+        map_c_h = false,
+        map_c_w = false,
+      })
+    end
   end,
 }
