+++ lua/plugins/diffview.lua
@@ -1,254 +1,5 @@
--- Code quality & analysis: metrics, security, profiling, refactoring, dependencies
+-- Advanced git analysis
 return {
-  {
-    -- Code metrics and complexity analysis
-    "mfussenegger/nvim-lint",
-    config = function()
-      require("lint").linters_by_ft = {
-        python = { "flake8", "mypy", "bandit" },
-        javascript = { "eslint" },
-        typescript = { "eslint" },
-        go = { "golangcilint", "staticcheck" },
-        terraform = { "tflint", "tfsec" },
-        yaml = { "yamllint" },
-        json = { "jsonlint" },
-        lua = { "luacheck" },
-        bash = { "shellcheck" },
-        elixir = { "credo" },
-      }
-
-      -- Auto-run linting
-      vim.api.nvim_create_autocmd({ "BufWritePost", "BufReadPost", "InsertLeave" }, {
-        callback = function()
-          require("lint").try_lint()
-        end,
-      })
-
-      -- Keymaps
-      vim.keymap.set("n", "<leader>cl", function()
-        require("lint").try_lint()
-      end, { desc = "Run linting" })
-    end,
-  },
-
-  {
-    -- Performance profiling for Lua
-    "stevearc/profile.nvim",
-    config = function()
-      local should_profile = os.getenv("NVIM_PROFILE")
-      if should_profile then
-        require("profile").instrument_autocmds()
-        if should_profile:lower():match("^start") then
-          require("profile").start("*")
-        else
-          require("profile").instrument("*")
-        end
-      end
-
-      local function toggle_profile()
-        local prof = require("profile")
-        if prof.is_recording() then
-          prof.stop()
-          vim.ui.input(
-            { prompt = "Save profile to:", completion = "file", default = "profile.json" },
-            function(filename)
-              if filename then
-                prof.export(filename)
-                vim.notify(string.format("Wrote %s", filename))
-              end
-            end
-          )
-        else
-          prof.start("*")
-        end
-      end
-
-      vim.keymap.set("n", "<leader>pp", toggle_profile, { desc = "Toggle profiling" })
-    end,
-  },
-  {
-    -- Advanced refactoring tools
-    "ThePrimeagen/refactoring.nvim",
-    dependencies = {
-      "nvim-lua/plenary.nvim",
-      "nvim-treesitter/nvim-treesitter",
-    },
-    config = function()
-      require("refactoring").setup({
-        prompt_func_return_type = {
-          go = false,
-          java = false,
-          cpp = false,
-          c = false,
-          h = false,
-          hpp = false,
-          cxx = false,
-        },
-        prompt_func_param_type = {
-          go = false,
-          java = false,
-          cpp = false,
-          c = false,
-          h = false,
-          hpp = false,
-          cxx = false,
-        },
-        printf_statements = {},
-        print_var_statements = {},
-        show_success_message = true,
-      })
-
-      -- Keymaps
-      vim.keymap.set({ "n", "x" }, "<leader>rr", function()
-        require("refactoring").select_refactor()
-      end, { desc = "Refactoring menu" })
-
-      -- Extract function supports only visual mode
-      vim.keymap.set("x", "<leader>re", ":Refactor extract ", { desc = "Extract function" })
-      vim.keymap.set("x", "<leader>rf", ":Refactor extract_to_file ", { desc = "Extract to file" })
-
-      -- Extract variable supports only visual mode
-      vim.keymap.set("x", "<leader>rv", ":Refactor extract_var ", { desc = "Extract variable" })
-
-      -- Inline func supports only normal
-      vim.keymap.set("n", "<leader>rI", ":Refactor inline_func", { desc = "Inline function" })
-
-      -- Inline var supports both normal and visual mode
-      vim.keymap.set({ "n", "x" }, "<leader>ri", ":Refactor inline_var", { desc = "Inline variable" })
-
-      -- Extract block supports only normal mode
-      vim.keymap.set("n", "<leader>rb", ":Refactor extract_block", { desc = "Extract block" })
-      vim.keymap.set("n", "<leader>rbf", ":Refactor extract_block_to_file", { desc = "Extract block to file" })
-
-      -- Debug helpers
-      vim.keymap.set({ "n", "x" }, "<leader>rp", function()
-        require("refactoring").debug.printf({ below = false })
-      end, { desc = "Debug print" })
-      vim.keymap.set("n", "<leader>rpv", function()
-        require("refactoring").debug.print_var({ normal = true })
-      end, { desc = "Debug print variable" })
-      vim.keymap.set("x", "<leader>rpv", function()
-        require("refactoring").debug.print_var({})
-      end, { desc = "Debug print variable" })
-      vim.keymap.set("n", "<leader>rpc", function()
-        require("refactoring").debug.cleanup({})
-      end, { desc = "Debug cleanup" })
-    end,
-  },
-  {
-    -- Dependency analysis
-    "vuki656/package-info.nvim",
-    dependencies = { "MunifTanjim/nui.nvim" },
-    config = function()
-      require("package-info").setup({
-        highlight = {
-          up_to_date = "#3C4048",
-          outdated = "#d19a66",
-        },
-        icons = {
-          enable = true,
-          style = {
-            up_to_date = "|  ",
-            outdated = "|  ",
-          },
-        },
-        autostart = true,
-        hide_up_to_date = false,
-        hide_unstable_versions = false,
-        package_manager = "npm",
-      })
-
-      -- Keymaps
-      vim.keymap.set(
-        "n",
-        "<leader>ns",
-        require("package-info").show,
-        { desc = "Show dependency versions", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>nc",
-        require("package-info").hide,
-        { desc = "Hide dependency versions", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>nt",
-        require("package-info").toggle,
-        { desc = "Toggle dependency versions", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>nu",
-        require("package-info").update,
-        { desc = "Update dependency on line", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>nd",
-        require("package-info").delete,
-        { desc = "Delete dependency on line", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>ni",
-        require("package-info").install,
-        { desc = "Install new dependency", silent = true }
-      )
-      vim.keymap.set(
-        "n",
-        "<leader>np",
-        require("package-info").change_version,
-        { desc = "Change dependency version", silent = true }
-      )
-    end,
-  },
-  {
-    -- Code coverage heatmap
-    "andythigpen/nvim-coverage",
-    dependencies = { "nvim-lua/plenary.nvim" },
-    config = function()
-      require("coverage").setup({
-        commands = true,
-        highlights = {
-          covered = { fg = "#C3E88D" },
-          uncovered = { fg = "#F07178" },
-          partial = { fg = "#AA71FF" },
-        },
-        signs = {
-          covered = { hl = "CoverageCovered", text = "▎" },
-          uncovered = { hl = "CoverageUncovered", text = "▎" },
-          partial = { hl = "CoveragePartial", text = "▎" },
-        },
-        summary = {
-          min_coverage = 80.0,
-        },
-        lang = {
-          python = {
-            coverage_command = "coverage json --fail-under=0 -q -o -",
-            coverage_file = "coverage.json",
-          },
-          go = {
-            coverage_command = "go test -coverprofile=coverage.out ./... && go tool cover -func=coverage.out",
-          },
-          javascript = {
-            coverage_command = "cat coverage/lcov-report/index.html | grep -o '\\\"decimal\\\">[^<]*' | head -n 4",
-          },
-        },
-        auto_reload = true,
-        load_coverage_cb = function(ftype)
-          vim.notify("Loaded " .. ftype .. " coverage")
-        end,
-      })
-
-      -- Keymaps
-      vim.keymap.set("n", "<leader>cov", "<cmd>Coverage<cr>", { desc = "Show coverage" })
-      vim.keymap.set("n", "<leader>coh", "<cmd>CoverageHide<cr>", { desc = "Hide coverage" })
-      vim.keymap.set("n", "<leader>cos", "<cmd>CoverageSummary<cr>", { desc = "Coverage summary" })
-      vim.keymap.set("n", "<leader>cot", "<cmd>CoverageToggle<cr>", { desc = "Toggle coverage" })
-      vim.keymap.set("n", "<leader>col", "<cmd>CoverageLoad<cr>", { desc = "Load coverage" })
-    end,
-  },
   {
     -- Advanced git analysis
     "sindrets/diffview.nvim",
