+++ lua/plugins/emojis-nvim.lua
@@ -1,0 +1,19 @@
+return {
+  "allaman/emoji.nvim",
+  version = "5.0.0",
+  ft = "markdown",
+  dependencies = {
+    "nvim-lua/plenary.nvim",
+    "saghen/blink.cmp",
+    "nvim-telescope/telescope.nvim",
+    "ibhagwan/fzf-lua",
+  },
+  opts = {
+    enable_cmp_integration = true,
+  },
+  config = function(_, opts)
+    require("emoji").setup(opts)
+    local ts = require("telescope").load_extension("emoji")
+    vim.keymap.set("n", "<leader>se", ts.emoji, { desc = "[S]earch [E]moji" })
+  end,
+}
