+++ lua/plugins/lsp-signature.lua
@@ -1,0 +1,47 @@
+-- Enhanced signature help
+return {
+  {
+    -- Enhanced signature help
+    "ray-x/lsp_signature.nvim",
+    config = function()
+      require("lsp_signature").setup({
+        bind = true,
+        doc_lines = 10,
+        max_height = 12,
+        max_width = 80,
+        wrap = true,
+        floating_window = true,
+        floating_window_above_cur_line = true,
+        floating_window_off_x = 1,
+        floating_window_off_y = 0,
+        close_timeout = 4000,
+        fix_pos = false,
+        hint_enable = true,
+        hint_prefix = "🐼 ",
+        hint_scheme = "String",
+        hi_parameter = "LspSignatureActiveParameter",
+        handler_opts = {
+          border = "rounded",
+        },
+        always_trigger = false,
+        auto_close_after = nil,
+        extra_trigger_chars = {},
+        zindex = 200,
+        padding = " ",
+        transparency = nil,
+        shadow_blend = 36,
+        shadow_guibg = "Black",
+        timer_interval = 200,
+        toggle_key = nil,
+        select_signature_key = nil,
+        move_cursor_key = nil,
+        -- Filter out terminal buffers to prevent URI errors
+        filter = function(bufnr)
+          local buftype = vim.bo[bufnr].buftype
+          local filetype = vim.bo[bufnr].filetype
+          return buftype ~= "terminal" and filetype ~= "toggleterm"
+        end,
+      })
+    end,
+  },
+}
