+++ lua/plugins/lspconfig.lua
@@ -1,22 +1,11 @@
 return {
   "neovim/nvim-lspconfig",
   dependencies = {
-    "hrsh7th/nvim-cmp",
-    "hrsh7th/cmp-buffer",
-    "hrsh7th/cmp-path",
-    "saadparwaiz1/cmp_luasnip",
-    "hrsh7th/cmp-nvim-lsp",
-    "L3MON4D3/LuaSnip",
-    "rafamadriz/friendly-snippets",
+    "saghen/blink.cmp",
   },
   config = function()
     local lspconfig = require("lspconfig")
-    local lsp_capabilities = require("cmp_nvim_lsp").default_capabilities()
-    require("luasnip.loaders.from_vscode").lazy_load()
-    vim.opt.completeopt = { "menu", "menuone", "noselect" }
-    local cmp = require("cmp")
-    local luasnip = require("luasnip")
-    local select_opts = { behavior = cmp.SelectBehavior.Insert }
+    local lsp_capabilities = require("blink.cmp").get_lsp_capabilities()
     -- Configure diagnostic signs (modern approach)
     vim.diagnostic.config({
       signs = {
@@ -57,84 +46,6 @@ return {
 
     vim.lsp.handlers["textDocument/signatureHelp"] =
       vim.lsp.with(vim.lsp.handlers.signature_help, { border = "rounded" })
-
-    cmp.setup({
-      experimental = {
-        ghost_text = true,
-      },
-      snippet = {
-        expand = function(args)
-          luasnip.lsp_expand(args.body)
-        end,
-      },
-      sources = {
-        { name = "nvim_lsp", priority = 1000 },
-        { name = "luasnip", priority = 750 },
-        { name = "path", priority = 500 },
-        { name = "buffer", keyword_length = 3, priority = 250 },
-      },
-      window = {
-        documentation = {
-          border = "rounded",
-          cmp.config.window.bordered(),
-        },
-      },
-      formatting = {
-        fields = { "menu", "abbr", "kind" },
-        format = function(entry, item)
-          local menu_icon = {
-            nvim_lsp = "λ",
-            luasnip = "⋗",
-            buffer = "Ω",
-            path = "🖫",
-          }
-
-          item.menu = menu_icon[entry.source.name]
-          return item
-        end,
-      },
-      mapping = {
-        ["<C-p>"] = cmp.mapping.select_prev_item(select_opts),
-        ["<C-n>"] = cmp.mapping.select_next_item(select_opts),
-        ["<C-u>"] = cmp.mapping.scroll_docs(-4),
-        ["<C-d>"] = cmp.mapping.scroll_docs(4),
-        ["<C-e>"] = cmp.mapping.abort(),
-        ["<C-y>"] = cmp.mapping.confirm({ select = true }),
-        ["<CR>"] = cmp.mapping.confirm({ select = false }),
-        ["<C-f>"] = cmp.mapping(function(fallback)
-          if luasnip.jumpable(1) then
-            luasnip.jump(1)
-          else
-            fallback()
-          end
-        end, { "i", "s" }),
-        ["<C-b>"] = cmp.mapping(function(fallback)
-          if luasnip.jumpable(-1) then
-            luasnip.jump(-1)
-          else
-            fallback()
-          end
-        end, { "i", "s" }),
-        ["<Tab>"] = cmp.mapping(function(fallback)
-          local col = vim.fn.col(".") - 1
-
-          if cmp.visible() then
-            cmp.select_next_item(select_opts)
-          elseif col == 0 or vim.fn.getline("."):sub(col, col):match("%s") then
-            fallback()
-          else
-            cmp.complete()
-          end
-        end, { "i", "s" }),
-        ["<S-Tab>"] = cmp.mapping(function(fallback)
-          if cmp.visible() then
-            cmp.select_prev_item(select_opts)
-          else
-            fallback()
-          end
-        end, { "i", "s" }),
-      },
-    })
 
     -- LSP servers are configured via mason-lspconfig handlers
     -- No need for manual setup here
