+++ lua/plugins/mason-lspconfig.lua
@@ -15,19 +15,59 @@ return {
       },
       handlers = {
         function(server_name)
-          local capabilities = require("cmp_nvim_lsp").default_capabilities()
+          local capabilities = require("blink.cmp").get_lsp_capabilities()
           require("lspconfig")[server_name].setup({
             capabilities = capabilities,
           })
         end,
         ["terraformls"] = function()
           require("lspconfig").terraformls.setup({
-            capabilities = require("cmp_nvim_lsp").default_capabilities(),
+            capabilities = require("blink.cmp").get_lsp_capabilities(),
             settings = {
               terraformls = {
                 ignoreSingleFileWarning = true,
+                experimentalFeatures = {
+                  validateOnSave = true,
+                  prefillRequiredFields = true,
+                  moduleVariableHints = true,
+                },
+                indexing = {
+                  enabled = true,
+                  ignorePaths = {
+                    ".terraform/providers/**",
+                    ".terraform/.terraform.lock.hcl",
+                  },
+                  modulePaths = {
+                    ".terraform/modules/**",
+                    "./modules/**",
+                    "../modules/**",
+                    "../../modules/**",
+                    "./terraform-modules/**",
+                    "../terraform-modules/**",
+                  },
+                },
+                validation = {
+                  enableEnhancedValidation = true,
+                },
               },
             },
+            on_attach = function(client, bufnr)
+              -- Enable inlay hints for module variables
+              if client.server_capabilities.inlayHintProvider then
+                vim.lsp.inlay_hint.enable(true, { bufnr = bufnr })
+              end
+
+              -- Add keymap to show module info
+              vim.keymap.set("n", "<leader>tV", function()
+                -- Get current word under cursor
+                local word = vim.fn.expand("<cword>")
+                if word and word ~= "" then
+                  vim.lsp.buf.hover()
+                else
+                  print("Place cursor on module name and try again")
+                end
+              end, { buffer = bufnr, desc = "Show Module Info" })
+            end,
           })
         end,
       },
