+++ lua/plugins/obsidian.lua
@@ -12,7 +12,7 @@ return {
   -- },
   dependencies = {
     "nvim-lua/plenary.nvim",
-    "hrsh7th/nvim-cmp",
+    "saghen/blink.cmp",
     "nvim-telescope/telescope.nvim",
     "nvim-treesitter/nvim-treesitter",
   },
