+++ lua/plugins/schemastore.lua
@@ -1,0 +1,10 @@
+-- JSON tools
+return {
+  {
+    -- JSON tools
+    "b0o/schemastore.nvim",
+    config = function()
+      -- This will be used by jsonls via lspconfig
+    end,
+  },
+}
