+++ lua/plugins/vim-terraform.lua
@@ -1,0 +1,11 @@
+-- Terraform tools
+return {
+  {
+    -- Terraform tools
+    "hashivim/vim-terraform",
+    config = function()
+      vim.g.terraform_align = 1
+      vim.g.terraform_fmt_on_save = 1
+    end,
+  },
+}
